Working deposit and withdrawl

+++ BankingManagement/src/main/java/BankingManagement/service/BankingService.java
@@ -7,9 +7,9 @@ public interface BankingService {
 	
 	public Banking createBank(Banking bank);
 	
-	public void deposit(Banking bank, int amount);
+	public void deposit(Banking bank, double amount);
 	
-	public void withdrawl(Banking bank, int amount);
+	public void withdrawl(Banking bank, double amount);
 	
 	public double retrieveBalance(Banking bank);
 }
